requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-core
sentence-transformers
chromadb
python-dotenv
fastapi
uvicorn
pydantic
streamlit